Refactored the rotation search code and moved all the scoring classes into 1 directory
