+++ simbad/score/phaser_score.py
@@ -4,20 +4,22 @@ __author__ = "Adam Simpkin"
 __date__ = "27 Dec 2017"
 __version__ = "0.1"
 
+from simbad.score import ScoreBase
 
-class PhaserRotationScore(object):
+
+class PhaserRotationScore(ScoreBase):
     """An phaser rotation scoring class"""
 
-    __slots__ = ("pdb_code", "dat_path", "LLG", "RFZ")
+    __slots__ = ("pdb_code", "dat_path", "llg", "rfz")
 
-    def __init__(self, pdb_code, dat_path, LLG, RFZ):
+    def __init__(self, pdb_code, dat_path, llg, rfz):
         self.pdb_code = pdb_code
         self.dat_path = dat_path
-        self.LLG = LLG
-        self.RFZ = RFZ
+        self.llg = llg
+        self.rfz = rfz
 
     def __repr__(self):
-        string = "{name}(pdb_code={pdb_code} dat_path={dat_path} LLG={LLG} RFZ={RFZ})"
+        string = "{name}(pdb_code={pdb_code} dat_path={dat_path} llg={llg} rfz={rfz})"
         return string.format(name=self.__class__.__name__, **{k: getattr(self, k) for k in self.__slots__})
 
     def _as_dict(self):
